Add files via upload
MemoryLeakTest.cpp is the code for 3D PACMAN with CPP array instead of Eigen. At the moment the code runs till iteration 19 and blows up which suggests an error internally. The plots look correct, however, are ''flipped'' or could be indexing problem. So, this needs to be tested to make sure indexing is okay. For now will keep this here and will change the local copy to include new FFT functions with separate plan creation and faster executes.